backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-docx==1.1.2
sqlmodel==0.0.22
python-multipart==0.0.9
